Separation of Distillation Column Project (CHE411)
Mass Transfer Operations final project. Software was used to automate the calculation of equilibrium stages within the column. Equilibrium data of water-ethanol interaction is found in the excel file. The software was used to optimize the recycle ratio with respect to cost. Future work includes implementing/solving the initial balances and adding a cost minimization algorithm.